fix: Fix work with labelstudio offsets
