+++ convert/tests/labelstudio/export/test_export_empty_annotations/data_with_empty_annotations/badgerdoc/annotation/annotations_empty.json
