+++ convert/tests/labelstudio/export/test_export_empty_annotations/data_with_empty_annotations/badgerdoc/tokens/tokens_without_whitespaces.json
