+++ convert/tests/labelstudio/export/test_export_with_links/data_annotations_with_links/badgerdoc/annotation/fixed.json
@@ -1,0 +1,84 @@
+{
+  "size": {
+    "width": 79.0,
+    "height": 41.0
+  },
+  "page_num": 1,
+  "objs": [
+    {
+      "id": 0,
+      "type": "text",
+      "bbox": [
+        15.0,
+        15.0,
+        43.0,
+        26.0
+      ],
+      "category": "Jurisdiction_v23",
+      "data": {
+        "tokens": [
+          {
+            "id": 0,
+            "text": "1",
+            "x": 15.0,
+            "y": 15.0,
+            "width": 7.0,
+            "height": 11.0
+          },
+          {
+            "id": 1,
+            "text": "2 ",
+            "x": 22.0,
+            "y": 15.0,
+            "width": 7.0,
+            "height": 11.0
+          },
+          {
+            "id": 2,
+            "text": "4  ",
+            "x": 36.0,
+            "y": 15.0,
+            "width": 7.0,
+            "height": 11.0
+          }
+        ],
+        "dataAttributes": []
+      },
+      "text": "12 4",
+      "links": []
+    },
+    {
+      "id": 1,
+      "type": "text",
+      "bbox": [
+        57.0,
+        15.0,
+        64.0,
+        26.0
+      ],
+      "category": "Person_v23",
+      "data": {
+        "tokens": [
+          {
+            "id": 3,
+            "text": "7",
+            "x": 57.0,
+            "y": 15.0,
+            "width": 7.0,
+            "height": 11.0
+          }
+        ],
+        "dataAttributes": []
+      },
+      "text": "7",
+      "links": [
+        {
+          "category_id": "Link",
+          "to": 0,
+          "type": "directional",
+          "page_num": 1
+        }
+      ]
+    }
+  ]
+}
