+++ convert/tests/labelstudio/export/test_export_with_links/data_annotations_with_links/badgerdoc/annotation/manifest.json
@@ -1,0 +1,16 @@
+{
+  "revision": "1442f2166033e4fbf487cfdcc41c2eff86a0e8d7",
+  "job_id": 2190,
+  "user": null,
+  "pipeline": 0,
+  "date": "2023-04-05T17:14:17.778303",
+  "pages": {
+    "1": "fixed"
+  },
+  "failed_validation_pages": [],
+  "validated": [],
+  "links_json": [],
+  "file": "files/1065/1065.pdf",
+  "bucket": "test",
+  "categories": []
+}
