+++ convert/tests/labelstudio/import/test_import_annotation/data_annotations/badgerdoc/tokens/1.json
@@ -1,0 +1,67 @@
+{
+  "page_num": 1,
+  "size": {
+    "width": 79.0,
+    "height": 41.0
+  },
+  "objs": [
+    {
+      "type": "text",
+      "bbox": [
+        15.0,
+        15.0,
+        22.0,
+        26.0
+      ],
+      "text": "1",
+      "offset": {
+        "begin": 0,
+        "end": 1
+      }
+    },
+    {
+      "type": "text",
+      "bbox": [
+        22.0,
+        15.0,
+        29.0,
+        26.0
+      ],
+      "text": "2",
+      "offset": {
+        "begin": 1,
+        "end": 2
+      },
+      "after": " "
+    },
+    {
+      "type": "text",
+      "bbox": [
+        36.0,
+        15.0,
+        43.0,
+        26.0
+      ],
+      "text": "4",
+      "offset": {
+        "begin": 3,
+        "end": 4
+      },
+      "after": "  "
+    },
+    {
+      "type": "text",
+      "bbox": [
+        57.0,
+        15.0,
+        64.0,
+        26.0
+      ],
+      "text": "7",
+      "offset": {
+        "begin": 6,
+        "end": 7
+      }
+    }
+  ]
+}
