+++ convert/tests/labelstudio/import/test_import_document_links/data_annotations_with_links/labelstudio.json
@@ -1,0 +1,79 @@
+[
+  {
+    "id": 75,
+    "annotations": [
+      {
+        "id": 35,
+        "completed_by": 1,
+        "result": [
+          {
+            "value": {
+              "start": 0,
+              "end": 4,
+              "text": "12 4",
+              "labels": [
+                "Jurisdiction_v23"
+              ]
+            },
+            "id": "2BTg_zS4Wd",
+            "from_name": "label",
+            "to_name": "text",
+            "type": "labels",
+            "origin": "manual"
+          },
+          {
+            "value": {
+              "start": 6,
+              "end": 7,
+              "text": "7",
+              "labels": [
+                "Person_v23"
+              ]
+            },
+            "id": "tj_1pQVFG4",
+            "from_name": "label",
+            "to_name": "text",
+            "type": "labels",
+            "origin": "manual"
+          },
+          {
+            "from_id": "tj_1pQVFG4",
+            "to_id": "2BTg_zS4Wd",
+            "type": "relation",
+            "direction": "right",
+            "labels": []
+          }
+        ],
+        "was_cancelled": false,
+        "ground_truth": false,
+        "created_at": "2023-04-07T11:19:32.658393Z",
+        "updated_at": "2023-04-07T11:20:24.455881Z",
+        "lead_time": 127.254,
+        "prediction": {},
+        "result_count": 0,
+        "task": 75,
+        "parent_prediction": null,
+        "parent_annotation": null
+      }
+    ],
+    "file_upload": "cfde5517-project-labelstudio.json",
+    "drafts": [],
+    "predictions": [],
+    "data": {
+      "text": "12 4  7"
+    },
+    "meta": {},
+    "created_at": "2023-04-07T11:19:32.657268Z",
+    "updated_at": "2023-04-07T11:20:24.485865Z",
+    "inner_id": 2,
+    "total_annotations": 1,
+    "cancelled_annotations": 0,
+    "total_predictions": 0,
+    "comment_count": 0,
+    "unresolved_comment_count": 0,
+    "last_comment_updated_at": null,
+    "project": 19,
+    "updated_by": 1,
+    "comment_authors": []
+  }
+]
