--- convert/tests/test_labelstudio/test_data/badgerdoc_etalon/annotations_test.json
@@ -1,148 +1,0 @@
-{
-  "size": {
-    "width": 590,
-    "height": 63
-  },
-  "page_num": 1,
-  "objs": [
-    {
-      "id": 0,
-      "type": "text",
-      "bbox": [
-        344,
-        15,
-        407,
-        26
-      ],
-      "category": "Organization",
-      "data": {
-        "tokens": [
-          {
-            "id": 47,
-            "text": "J",
-            "x": 344,
-            "y": 15,
-            "width": 7,
-            "height": 11
-          },
-          {
-            "id": 48,
-            "text": "e",
-            "x": 351,
-            "y": 15,
-            "width": 7,
-            "height": 11
-          },
-          {
-            "id": 49,
-            "text": "r",
-            "x": 358,
-            "y": 15,
-            "width": 7,
-            "height": 11
-          },
-          {
-            "id": 50,
-            "text": "u",
-            "x": 365,
-            "y": 15,
-            "width": 7,
-            "height": 11
-          },
-          {
-            "id": 51,
-            "text": "s",
-            "x": 372,
-            "y": 15,
-            "width": 7,
-            "height": 11
-          },
-          {
-            "id": 52,
-            "text": "a",
-            "x": 379,
-            "y": 15,
-            "width": 7,
-            "height": 11
-          },
-          {
-            "id": 53,
-            "text": "l",
-            "x": 386,
-            "y": 15,
-            "width": 7,
-            "height": 11
-          },
-          {
-            "id": 54,
-            "text": "e",
-            "x": 393,
-            "y": 15,
-            "width": 7,
-            "height": 11
-          },
-          {
-            "id": 55,
-            "text": "m",
-            "x": 400,
-            "y": 15,
-            "width": 7,
-            "height": 11
-          }
-        ],
-        "dataAttributes": []
-      },
-      "text": "Jerusalem",
-      "links": [
-        {
-          "category_id": "some_relation",
-          "to": 1,
-          "type": "directional",
-          "page_num": 1
-        }
-      ]
-    },
-    {
-      "id": 1,
-      "type": "text",
-      "bbox": [
-        218,
-        15,
-        239,
-        26
-      ],
-      "category": "Datetime",
-      "data": {
-        "tokens": [
-          {
-            "id": 29,
-            "text": "2",
-            "x": 218,
-            "y": 15,
-            "width": 7,
-            "height": 11
-          },
-          {
-            "id": 30,
-            "text": "7",
-            "x": 225,
-            "y": 15,
-            "width": 7,
-            "height": 11
-          },
-          {
-            "id": 31,
-            "text": " ",
-            "x": 232,
-            "y": 15,
-            "width": 7,
-            "height": 11
-          }
-        ],
-        "dataAttributes": []
-      },
-      "text": "27 ",
-      "links": []
-    }
-  ]
-}
