+++ convert/tests/text/test_remove_non_printing_tokens.py
